open the listening socket before returning from groupcache_create()
creating the socket earlier, instead of doing that directly in the background thread,
ensures being ready to serve new connections as soon as we are back from groupcache_create().
So if a script/program creates a new groupcaches instance and queries it with a client immediately after
groupcache_create() returned, it won't fail connecting because there is no listening socket yet.