updated libiomux
major fix for epoll/kqueue implementation, it was leaking the epoll/kqueue fd
when the mux was destroyed. It was unnoticed so far because we were allocating
one mux per workers which would have lived until the end of the program.
Now that the asynchronous readers create iomuxes on the fly the fd leakage was
immediately noticed