removed the logic which took care of emptying the read buffer  before giving a filedescriptor back to the caller
it wasn't reliable anyway since data might be still in transit and not available
and adding more overhead putting the filedescriptors in quarantine before reusing
is worthless.

In the end a caller must ensure putting back filedescriptors into the cache only
when really finished with them ... or anyway be resilient to receiving data that
the previous user didn't read (on purpose).