don't insist sending an evict command to a dead peer
retry 5 times, but if still we can't connect we need to abandon and move forwards.
Note that there is no need to 'sleep' before retrying because we already have a connect timeout.
If the host is really dead (and we get a reset back) there is not even point in trying to send the
evict command since most likely the process is dead and needs to be restarted (hence the cache will be empty already)