use the macro-based linked list instead of the thread-safe one
since we don't need it to be thread safe and it's going to be used very frequently
so calls to pthread_mutex_lock() start to matter.

Also updated libiomux because of some optimizations there as well list instead of the thread-safe one

since we don't need it to be thread safe and it's going to be used very frequently
so calls to pthread_mutex_lock() start to matter.

Also updated libiomux because of some optimizations there as well list instead of the thread-safe one

since we don't need it to be thread safe and it's going to be used very frequently
so calls to pthread_mutex_lock() start to matter.

Also updated libiomux because of some optimizations there as well list instead of the thread-safe one

since we don't need it to be thread safe and it's going to be used very frequently
so calls to pthread_mutex_lock() start to matter.

Also updated libiomux because of some optimizations there as well